reverted dtype change

+++ test/test_ops.py
@@ -1133,7 +1133,7 @@ class TestOps(unittest.TestCase):
     c = torch.randint(low=-5, high=5, size=(10,1,14,1,1,1), requires_grad=False)
     d = torch.randint(high=4, size=(10,1,1,14,1,1), requires_grad=False)
     e = torch.randint(high=1, size=(10,1,1,1,13,1), requires_grad=False)
-    i, j, k, o, p = [Tensor(tor.numpy(), requires_grad=False).float() for tor in [a,b,c,d,e]]
+    i, j, k, o, p = [Tensor(tor.numpy(), requires_grad=False) for tor in [a,b,c,d,e]]
     helper_test_op([(2,5,15,5,3,4)], lambda x: x[a,b,c,d,e], lambda x: x[i,j,k,o,p])
     helper_test_op([(2,5,15,5,3,4)], lambda x: x[:,b,c,d,e], lambda x: x[:,j,k,o,p])
     helper_test_op([(2,5,15,5,3,4)], lambda x: x[:,b,c,d,:], lambda x: x[:,j,k,o,:])
@@ -1149,7 +1149,7 @@ class TestOps(unittest.TestCase):
   def test_gather(self):
     # indices cannot have gradient
     c = np.random.randint(3, size=[3,4,5,1,4]).astype(np.int32)
-    a = Tensor(c, requires_grad=False).float()
+    a = Tensor(c, requires_grad=False)
     b = torch.tensor(c, dtype=torch.int64, requires_grad=False)
     helper_test_op([(4,5,6,9,5)], lambda x: x.gather(index=b, dim=0), lambda x: x.gather(idx=a, dim=0))
     helper_test_op([(4,5,6,9,5)], lambda x: x.gather(index=b, dim=1), lambda x: x.gather(idx=a, dim=1))
